!!!Add hmac verification for index matcher
This prevents security issues when the matcher is loaded from the index file
and not provided in the `openIndex` or `ensureIndex` calls.
The hmac of the matcher calculated with the key provided during Storage
creation will be compared with the hmac retrieved from the index metadata.
Also a safeguard is put in place to prevent accidentially creating invalid
primary indexes by providing a `matcher` option in the indexOptions.

NOTE: This change breaks all previously generated secondary indexes. You can
safely delete those indexes if you open them with `ensureIndex`. Otherwise
rebuild them once with `createIndex`.